# api/openapi/core-api.yaml
openapi: 3.0.3
info:
  title: Mint Core API
  version: 1.0.0
  description: |
    Public REST contract exposed by the `core` service for desktop/web frontend clients.
    Write operations are asynchronous (202 Accepted + command_id/correlation_id).
servers:
  - url: http://localhost:8091
    description: Local dev
...
paths:
...
  /api/v1/workspaces:
    post:
      tags: [Workspace]
      operationId: createWorkspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XCorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "202":
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateWorkspaceAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/workspaces/{workspace_id}:
    get:
      tags: [Workspace]
      operationId: getWorkspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Workspace projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/workspaces/{workspace_id}/channels:
    post:
      tags: [Workspace]
      operationId: createChannel
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/XCorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelRequest"
      responses:
        "202":
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateChannelAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/workspaces/{workspace_id}/channels/{channel_id}:
    get:
      tags: [Workspace]
      operationId: getChannel
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/ChannelID"
      responses:
        "200":
          description: Channel projection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelView"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/workspaces/{workspace_id}/members/{user_id}/join:
    post:
      tags: [Workspace]
      operationId: joinWorkspace
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
        - $ref: "#/components/parameters/UserID"
        - $ref: "#/components/parameters/XCorrelationID"
      responses:
        "202":
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinWorkspaceAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XCorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Optional correlation id propagated into command meta.
    WorkspaceID:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
    ChannelID:
      name: channel_id
      in: path
      required: true
      schema:
        type: string
    UserID:
      name: user_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CommandAccepted:
      type: object
      required: [command_id, correlation_id]
      properties:
        command_id:
          type: string
        correlation_id:
          type: string
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    CreateWorkspaceAccepted:
      allOf:
        - $ref: "#/components/schemas/CommandAccepted"
        - type: object
          required: [workspace_id]
          properties:
            workspace_id:
              type: string
    CreateChannelRequest:
      type: object
      required: [name, kind]
      properties:
        name:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [text, voice]
    CreateChannelAccepted:
      allOf:
        - $ref: "#/components/schemas/CommandAccepted"
        - type: object
          required: [workspace_id, channel_id]
          properties:
            workspace_id:
              type: string
            channel_id:
              type: string
    JoinWorkspaceAccepted:
      allOf:
        - $ref: "#/components/schemas/CommandAccepted"
        - type: object
          required: [workspace_id, user_id]
          properties:
            workspace_id:
              type: string
            user_id:
              type: string
    WorkspaceView:
      type: object
      required: [id, name, owner_id, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        owner_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ChannelView:
      type: object
      required: [workspace_id, id, name, kind, created_at, updated_at]
      properties:
        workspace_id:
          type: string
        id:
          type: string
        name:
          type: string
        kind:
          type: string
          enum: [text, voice]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...